+++ SourceCode/Services/Extensions/MeetingExtensions.cs
@@ -1,32 +1,35 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
 namespace ModulesRegistry.Services.Extensions;
-
-public static class LayoutExtensions
-{
-    public static string RegistrationOpensDate(this Layout layout) => layout.RegistrationOpeningDate.ToShortDateString();
-    public static string RegistrationClosesDate(this Layout layout) => layout.RegistrationClosingDate.ToShortDateString();
-    public static string RegistrationOfModulesClosesDate(this Layout layout) => (layout.ModuleRegistrationClosingDate ?? layout.RegistrationClosingDate).ToShortDateString();
-    internal static bool IsOpenForRegistration(this Layout layout, DateTime at) =>
-        layout.IsRegistrationPermitted &&
-        layout.RegistrationOpeningDate <= at &&
-        layout.RegistrationClosingDate >= at;
-
-    internal static bool IsNotYetOpenForRegistration(this Layout layout, DateTime at) =>
-        layout.IsRegistrationPermitted &&
-        layout.RegistrationOpeningDate > at;
-
-}
 public static class MeetingExtensions
 {
-    public static int DaysCount(this Meeting meeting) =>
-         (meeting.EndDate - meeting.StartDate).Days + 1;
-
     public static string Day(this Meeting meeting, int day) =>
         meeting.StartDate.AddDays(day - 1).DayOfWeek.ToString();
 
-    public static bool IsCancelled(this Meeting? meeting) =>
-        meeting is null || meeting.Status == (int)MeetingStatus.Canceled;
+    public static int DaysCount(this Meeting meeting) =>
+         (meeting.EndDate - meeting.StartDate).Days + 1;
+
+    public static bool IsCancelled([NotNullWhen(true)] this Meeting? meeting) =>
+        meeting is not null && meeting.Status == (int)MeetingStatus.Canceled;
+
+    private static bool IsNotCancelled([NotNullWhen(true)] this Meeting? meeting) =>
+        !meeting.IsCancelled();
+
+    public static bool IsOpenForRegistration([NotNullWhen(true)] this Meeting? meeting, DateTime at) =>
+        meeting is not null &&
+        meeting.IsNotCancelled() &&
+        meeting.IsAnyLayoutRegistrationPermitted() &&
+        meeting.Layouts.Any(l => l.IsOpenForRegistration(at));
+
+    private static bool IsClosedForRegistration([NotNullWhen(false)] this Meeting? meeting, DateTime at) =>
+        meeting is not null && 
+        meeting.IsAnyLayoutRegistrationPermitted() &&
+        meeting.Layouts.Any() && meeting.Layouts.All(l => l.RegistrationClosingDate <= at);
+
+    public static bool MayDelete(this Meeting? meeting, ClaimsPrincipal principal) =>
+        meeting is not null &&
+        principal.IsCountryOrGlobalAdministrator() && 
+        meeting.Layouts.Sum(l => l.LayoutParticipants.Count) == 0;
 
     public static DateTime? RegistrationOpensDate(this Meeting? meeting) =>
         meeting is null || !meeting.IsAnyLayoutRegistrationPermitted() ? null :
@@ -42,35 +45,15 @@ public static class MeetingExtensions
 
     public static bool IsNotYetOpenForRegistration([NotNullWhen(true)] this Meeting? meeting, DateTime at) =>
         meeting is not null &&
-        meeting.Layouts.Any(l => l.IsNotYetOpenForRegistration(at));
-
-    public static bool IsOpenForRegistration([NotNullWhen(true)] this Meeting? meeting, DateTime at, ClaimsPrincipal? principal = null)
-    {       
-        return principal is not null && meeting is not null && meeting.IsAnyLayoutRegistrationPermitted() && 
-        (
-            principal.IsMeetingOrganiserOrAdministrator(meeting) 
-        ) ||
-        (   meeting is not null && !meeting.IsCancelled() &&
-            meeting.Layouts.Any(l => l.IsOpenForRegistration(at)));
-    }
-
-    public static bool IsClosedForRegistration([NotNullWhen(false)] this Meeting? meeting, DateTime at) =>
-        meeting is not null && meeting.IsAnyLayoutRegistrationPermitted() &&
-        meeting.Layouts.Any() && meeting.Layouts.All(l => l.RegistrationClosingDate <= at);
+        meeting.Layouts.All(l => l.IsNotYetOpenForRegistration(at));
 
     private static bool IsAnyLayoutRegistrationPermitted(this Meeting meeting) =>
         meeting.Layouts.Any(l => l.IsRegistrationPermitted);
-
-    public static bool MayRegister(this Meeting? meeting, DateTime at, ClaimsPrincipal? principal) =>
-       principal is not null && principal.IsAnyAdministrator() ||
-        (principal.IsAuthenticated() && meeting.IsOpenForRegistration(at, principal));
-
 
     public static string Organiser(this Meeting? meeting) =>
         meeting is null ? string.Empty :
         meeting.GroupDomainId.HasValue ? $"{meeting.OrganiserGroup.FullName}/{meeting.GroupDomain?.Name}" :
         $"{meeting.OrganiserGroup.FullName}";
-
 
     public static string Status(this Meeting? meeting, DateTime at) =>
         meeting is null ? string.Empty :
@@ -84,7 +67,7 @@ public static class MeetingExtensions
         meeting.IsOrganiserInternal ? "Internal" :
         ((MeetingStatus)meeting.Status).ToString();
 
-    public static string MeetingStatusCssClass(this Meeting meeting, DateTime at) => 
+    public static string MeetingStatusCssClass(this Meeting meeting, DateTime at) =>
         $"meeting {meeting.StatusResourceName(at).ToLowerInvariant()}";