+++ SourceCode/Services/Implementations/CargoService.cs
@@ -98,7 +98,7 @@ public sealed class CargoService
     {
         if (principal is null) return null;
         using var dbContext = Factory.CreateDbContext();
-        if (principal.IsAnyAdministrator()) return await dbContext.Cargos.FindAsync(id).ConfigureAwait(false);
+        if (principal.IsCountryOrGlobalAdministrator()) return await dbContext.Cargos.FindAsync(id).ConfigureAwait(false);
         return null;
     }