+++ SourceCode/Services/Implementations/GroupService.cs
@@ -10,12 +10,12 @@ public sealed class GroupService
         Factory = factory;
     }
 
-    public async Task<IEnumerable<ListboxItem>> ListboxItemsAsync(ClaimsPrincipal? principal, int? countryId)
+    public async Task<IEnumerable<ListboxItem>> ListboxItemsAsync(ClaimsPrincipal? principal, int? countryId, int? onlyGroupId = null)
     {
         using var dbContext = Factory.CreateDbContext();
-        var id = principal.IsAnyAdministrator() ? 0 : countryId ?? principal.CountryId();
+        var id = principal.IsCountryOrGlobalAdministrator() ? 0 : countryId ?? principal.CountryId();
         return await dbContext.Groups
-            .Where(g =>  g.CountryId == id || id == 0)
+            .Where(g =>  (g.CountryId == id || id == 0) && onlyGroupId > 0 && g.Id == onlyGroupId.Value)
             .OrderBy(g => g.FullName)
             .Select(g => new ListboxItem(g.Id, g.FullName))
             .ToReadOnlyListAsync();