+++ SourceCode/Services/Implementations/StationService.cs
@@ -107,7 +107,7 @@ public class StationService
             {
                 using var dbContext = Factory.CreateDbContext();
                 var isDataAdministrator = await dbContext.GroupMembers.AsNoTracking().AnyAsync(gm => gm.IsDataAdministrator && gm.GroupId == ownershipRef.GroupId && gm.PersonId == principal.PersonId()).ConfigureAwait(false);
-                if (isDataAdministrator || principal.IsAnyAdministrator())
+                if (isDataAdministrator || principal.IsCountryOrGlobalAdministrator())
                 {
                     return await AddOrUpdate(dbContext, principal, entity, moduleId).ConfigureAwait(false);
                 }